+++ utils/context.go
@@ -7,5 +7,4 @@ package utils
 type contextKey string
 
 const ExecutionIdContextKey contextKey = "execution_id"
-const PluginNameContextKey contextKey = "plugin"
-const BuildIdContextKey contextKey = "build_id"
+const PluginContextKey contextKey = "plugin"
